requirements.txt
mermaid
md-mermaid
Markdown<3.2
elasticsearch
mdtable
pyyaml